requirements.txt
flask-restplus==0.13.0
Flask-SQLAlchemy==2.1
Werkzeug==0.16.1
pytest==6.1.2
Flask-HTTPAuth==3.1.2
asn1crypto==0.24.0
cryptography==2.3
itsdangerous==0.24
passlib==1.7.1
validate_email==1.3
Flask-Cors==1.10.3